requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
google-generativeai==0.3.2
langchain==0.1.5
langchain-google-genai==0.0.6
langgraph==0.0.20
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
